Most stable version so far
By revriting blocks' code from start to end for the second time I finally understood how it should work and comunicate. This version is so far the best, the most stable and the clearest.
From now on, the only problem with classic gameplay itself is the game sometimes freezing. Next commits will aim to resolve this problem